+++ src/panel/script/ScriptItem.vue
@@ -149,7 +149,7 @@ const exportScript = () => {
         {
           name: t`包含数据导出`,
           onClick: close => {
-            void downloadExport({ should_strip_data: true });
+            void downloadExport({ should_strip_data: false });
             close();
           },
         },